requirements.txt
aiodns==3.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
attrs==25.3.0
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
ffmpeg-python==0.2.0
frozenlist==1.7.0
future==1.0.0
idna==3.10
imageio-ffmpeg==0.6.0
msgspec==0.18.6
multidict==6.4.4
numpy==2.3.0
propcache==0.3.2
py-cord==2.6.1
pycares==4.9.0
pycord==0.1.1
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.1.0
requests==2.32.4
urllib3==2.4.0
yarl==1.20.1
yt-dlp==2025.6.9
